+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/commander_colour.js
@@ -59,7 +59,14 @@ define(function () {
         // We ran out of colours
         return minionColour;
       }
-      return [factions[faction][count], [192, 192, 192]];
+
+      var guardianColour = [255, 255, 255];
+      var secondaryColour = [192, 192, 192];
+      if (_.isEmpty(_.xor(minionColour[0], guardianColour))) {
+        return [guardianColour, secondaryColour];
+      } else {
+        return [(factions[faction][count], secondaryColour)];
+      }
     },
   };
 });